### 2017-06-02 #119 @mwmclean
RefManageR
Provides tools for importing and working with bibliographic references. Greatly enhances the bibentry class in R by providing a class BibEntry which stores BibTeX and BibLaTeX references, and can be easily searched by any field, by date ranges, and by various formats for name lists using R's person class.

```
Package: RefManageR
Version: 0.14.3
Title: Straightforward 'BibTeX' and 'BibLaTeX' Bibliography Management
Authors@R: person(c("Mathew", "W."), "McLean", role = c("aut", "cre"),
    email = "mathew.w.mclean@gmail.com")
Author: Mathew W. McLean [aut, cre]
Maintainer: Mathew W. McLean <mathew.w.mclean@gmail.com>
Description: Provides tools for importing and working with bibliographic
    references. It greatly enhances the 'bibentry' class by providing a class
    'BibEntry' which stores 'BibTeX' and 'BibLaTeX' references, supports 'UTF-8'
    encoding, and can be easily searched by any field, by date ranges, and by
    various formats for name lists (author by last names, translator by full names,
    etc.). Entries can be updated, combined, sorted, printed in a number of styles,
    and exported. 'BibTeX' and 'BibLaTeX' '.bib' files can be read into 'R' and
    converted to 'BibEntry' objects. Interfaces to 'NCBI Entrez', 'CrossRef', and
    'Zotero' are provided for importing references and references can be created
    from locally stored 'PDF' files using 'Poppler'. Includes functions for citing
    and generating a bibliography with hyperlinks for documents prepared with
    'RMarkdown' or 'RHTML'.
License: GPL-2 | GPL-3 | BSD_3_clause + file LICENSE
    xml2,
    plyr,
    tools,
    httr,
    bibtex,
    lubridate (>= 1.5.0),
    methods
```

- Please indicate which category or categories from our [package fit policies]

Reproducibility because it provides functions as well as vignette examples of working together with knitr to include references in reproducible reports.  Data access because of the functions it provides for importing bibliographic information into R.

Anyone that wants to write reproducible reports or simply organize bibliographic data from a variety of sources.

No package I'm aware of provides all the functionality, but there is some overlap with other packages.
 Base R provides the `bibentry` class.  `RefManageR` provides a class `BibEntry` that extends `bibentry` in a number of ways, such as support for `BibLaTeX`; improved handling of name list fields (author, editor, etc.); improved handling of dates (using `lubridate`); additional options for sorting, formatting, and printing; and additional methods that make it easier to search, sort, and modify parts of the `BibEntry` object.  There is a small overlap with the packages `rentrez` and `rcrossref`, in that `RefManageR` can make use of the same APIs that those packages use to get references.  However, `RefManageR` parses the results differently and returns a `BibEntry` object.

Aside from the packages and version of `R` listed above in the DESCRIPTION, [poppler](https://poppler.freedesktop.org/) is needed if a user want to extract bibliographic information from PDF files on their machine.

- [x] Do you intend for this package to go on CRAN?  (Has already been on CRAN for a few years)

__Note: I already submitted to JOSS [here](https://github.com/openjournals/joss-reviews/issues/269)__

__Exceptions__: I did not follow your guidelines when naming the functions in my package since they were named long ago and have been on CRAN for a long time now (I used title case and no underscores).

@karthik @cboettig @sckott
