### 2017-06-11 #119 @noamross
### Editor checks:

---

#### Editor comments

Thank you for your submission, @mwmclean! We were actually in the midst of revisions of our Aims and Scope when you submitted this, and one of our updates was to explicitly add reference management as a component of our reproducibility category.

Below are the results of our automated tests.  I don't see anything here to prevent review going forward, but we will want these addressed before acceptance.  This also provides a reference for reviewers of areas they may want to pay attention to. If you make changes in response to these or have a written response, please note them on the thread here so reviewers know about updates. I am currently seeking reviewer.

```
>goodpractice::gp()
── GP RefManageR ─────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package
    code in general. 66% of code lines are covered by test
    cases.

    R/02BibOptions.R:119:NA
    R/02BibOptions.R:120:NA
    R/04InternalFunctions.R:16:NA
    R/04InternalFunctions.R:17:NA
    R/04InternalFunctions.R:18:NA
    ... and 1794 more lines

  ✖ write short and simple functions. These functions have
    high cyclomatic complexity:GetFormatFunctions (130),
    MakeBibLaTeX (97), Cite (83), MakeAuthorYear (77),
    toBibtex.BibEntry (63), ResolveBibLaTeXCrossRef (58).
  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added
    to the package when you perform `R CMD build` on it.
  ✖ add a "URL" field to DESCRIPTION. It helps users find
    information about your package online. If your package
    does not have a homepage, add an URL to GitHub, or the
    CRAN package package page.
  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and
    it is easier to read your code for them if you use
    '<-'.

    R/BibEntryAddOp.R:52:19
    R/ReadGS.R:63:20
    R/ReadPDFsSupport.R:130:20
    R/ReadPubMed.R:60:21
    R/rmdCite.R:208:17

  ✖ avoid long code lines, it is bad for readability.
    Also, many people prefer editor windows that are about
    80 characters wide. Try make your lines shorter than
    80 characters

    R/02BibOptions.R:111:1
    R/02BibOptions.R:133:1
    R/02BibOptions.R:143:1
    R/02BibOptions.R:144:1
    R/02BibOptions.R:145:1
    ... and 785 more lines

    vector, or a list, depending on the input data.
    Consider using vapply() instead.

    R/04InternalFunctions.R:13:11
    R/04InternalFunctions.R:56:11
    R/04InternalFunctions.R:375:13
    R/04InternalFunctions.R:385:15
    R/04InternalFunctions.R:386:45
    ... and 79 more lines

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import
    only the specific functions you need.
  ✖ fix this R CMD check NOTE: Foreign function call to a
    different package: .External("do_read_bib", ...,
    PACKAGE = "bibtex") See chapter ‘System and foreign
    language interfaces’ in the ‘Writing R Extensions’
    manual.
───────────────────────────────────
```

Running your tests, I get the following messages, plus several more of the locale-based warnings.
```
Skipped ------------------------------------------------------------------------
1. Reading one PDF file name (@test-readPDF.R#68) - Couldn't copy jstor.pdf

2. Recognizes JSTOR (@test-readPDF.R#77) - Couldn't download Poppler or JSTOR PDF

Warnings -----------------------------------------------------------------------
1. PrintBibliography when multiple BibEntry objects cited (@test-cites.R#131) - input string 8 is invalid in this locale

2. Creates a BibEntry object (@test-readPDF.R#54) - back-tracking limit reached in PCRE for element 1

3. Creates a BibEntry object (@test-readPDF.R#54) - back-tracking limit reached in PCRE for element 1

4. Creates a BibEntry object (@test-readPDF.R#54) - input string 1 is invalid in this locale
```

In addition, `devtools::spell_check()` yields these (in addition to many other non-dictionary words):

```
simplying           SearchBib.Rd:48
requrested          ReadPDFs.Rd:34
mutliple            c.BibEntry.Rd:18
invisibling         head.BibEntry.Rd:33
gerenated           RefManageR-package.Rd:46
exctract            SearchBib.Rd:17, sub-sub-.BibEntry.Rd:12
decribed            print.BibEntry.Rd:49
coereced            sub-subset-.BibEntry.Rd:15
```

---

Reviewers:
Due date:
