### 2017-06-13 #119 @mwmclean
Thanks for the comments.  Regarding the last two notes from `goodpractice::gp()`, fixing them depends on the version of the [`bibtex`](https://github.com/romainfrancois/bibtex) package on GitHub.  However, I've been unable to contact the maintainer, @romainfrancois, to submit the updated version of `bibtex` to CRAN, and hence can't get rid of the NOTE.
