### 2017-06-24 #119 @noamross
Reviewers assigned: @cboettig @AmeliaMN. Due date 2017-07-15
