### 2017-06-27 #119 @cboettig
No CONTRIBUTING document or section of the Readme.

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
These do not lterally appear in paper.md, though paper.md uses pandoc-style citations and includes a .bib, so

---

Overview: The RefManageR package is overall a well designed and carefully engineered package that already sees significant use as indicated by the download statistics.  (Indeed, I've used the package myself both as a dependency and for occasional use. The source code is easily readible, well documented, and has recently migrated to the `httr` / `curl` / `xml2` stack. My only real comments are some big picture conceptual things that are no doubt beyond the scope of changes for onboarding, but since I have no more immediate critique to offer I just put them down here as food for thought.  Overall great package and nicely done.

The package aims to provide functions that fall into at least three separate buckets, which from the perspective of a developer, might have made more sense for these to be separated out into more complete, separate module packages. That could streamline maintenance and make it easier for other developers to just depend on the parts they need.  More importantly, it could help the package provide somewhat more complete implementations of these separate areas, and enforce a bit better definition of a low-level interface that a developer could use to leverage any part.  This will make more sense in terms of specifics, which I discuss below.  However, I suspect breaking up the package is beyond the scope here and should not prevent onboarding, but merely as a suggestion to the author and something to keep in mind for future directions.

The separate modules, as I see them, are:

- BibLaTex utilities: define a native R S3 class (BibEntry) corresponding to the BibLaTeX format, along with the ability to parse and serialize these files (much as bibtex package does does for bibtex).  This is a well-defined feature set that could easily stand alone, and be useful to others wanting a more fully-featured bibtex class.  It might be useful to have these in a seperate BibLatex package that is pulled in by RefManageR.

- Query and format citations in RMarkdown, etc.  These are basically some methods and functions for the BibEntry class that are useful for authors formatting references from R, which is a somewhat narrower application than the more general-purpose idea of providing a somewhat richer native bibliographic format to R.  These are the main user-facing functions of RefManageR.  These are really well-designed functions, and a nice compliment to the BibEntry that RefManageR provides.

- Query various external bibliographic formats.  The package supports calls to several different platforms (CrossRef, Zotero, NCBI) to import bibliographic metadata.  This is a useful ability to be sure, but is easily isolated from the above functions.

Unfortunately, I think separating out these elements also highlights some possible limitations of this design:

- With regards to the first 'module' of tasks, BibLatex is not an ideal format for providing a richer metadata model than R's native bibentry (which is basically bibtex). It's not easy to distinguish a biblatex from a bibtex file, and most academic publishers do not provide or accept BibLatex format. The similarity to vanilla bibtex seems likely to create a source of confusion for users who may fail to notice that biblatex files are not always bibtex compatible.

- A consequence of this choice impacts the third 'module' of tasks, querying external bibliographic providers.  These queries all seem to request data in `bibtex` format, which of course simplifies the programming, but also loses much of the data that could be available from the richer metadata formats (Something that is not obvious from the package documentation).  For instance, CrossRef/DataCite can tell us that a citation is a piece of software rather than an article or "misc", but this information is lost when we query their API by requesting bibtex.  This also defeats the advantage of having a native R format that was any richer than bibtex to begin with.

To me, these seem a bit like missed opportunities in the design.  The core "RefManageR" features in section 2 are excellent, but are restricted to only manipulating data in the biblatex format, and only accessing remote data in the bibtex format, which diminishes the potential somewhat.  The lack of more modular abstraction also encumbers the source code a little: for instance, this package makes no use of the existing (ropensci) rcrossref implementation, while it's native implementation makes rather heavy use of `try` calls for error handling, rather than the native http error-code handling apready built into `htt` (e.g. see  `httr::stop_for_status()`)
