### 2017-06-28 #119 @mwmclean
Thank you, @cboettig, for the quick and helpful report.  I will respond more completely to your report very soon, but for now just wanted to mention a couple updates and sticking points.

* Regarding the `goodpractice` checks, I now have the majority of them corrected.  The exceptions are
  * the last two already mentioned that are dependent on @romainfrancois submitting `bibtex` to CRAN
  * the note about cyclomatic complexity.  The first three functions listed define several small functions used for formatting entries for printing.  These functions are then returned in an environment that gets used when formatting, so I feel like these are false positives.  Will have a closer look at the last two functions listed.
* I have corrected the spelling mistakes in the documentation
* I am unable to reproduce the `warnings` mentioned from running the tests regarding "back-tracking limit reached in PCRE"
* I have just added a `CONTRIBUTING.md` file
* I thought I had included references properly using `paper.bib`, please let me know if I need to change it or `paper.md` in some way
