### 2017-06-30 #119 @romainfrancois
@mwmclean I'm about to upload `bibtex` 0.4.1 (what is on GitHub now)
