### 2017-07-17 #119 @AmeliaMN
The images in the `TestHTML` vignette aren't appearing for me. Otherwise, the vignettes look good. Related to @cboettig's comments about the three purposes of this package, the vignettes seem to cover mostly the second task. It might be useful to have a vignette illustrating the external querying functionality.

- [x] **Function Documentation:** for all exported functions in R helper

I noticed a possible spelling error in the title for the documentation on `GetPubMedByID()`. I think it should say E-Utilities rather than Entrez.

The example for `GetBibEntryWithDOI()` includes the function `http_error()`, which is in `httr`, and not automatically loaded by RefManageR.  A variety of other examples use this function (`GetDOIs`, `GetPubMedByID`, `GetPubMedRelated`, `LookupPubMed`, etc) so it might be worthwhile to import the function, or include the library call in the example.

When I edited the code to include `httr::http_error()`, I was able to get the example to run, but I got several warning messages:

```
1: In parse_Rd(Rd, encoding = encoding, fragment = fragment, ...) :
  <connection>:7: unexpected END_OF_INPUT '
'
2: In parse_Rd(Rd, encoding = encoding, fragment = fragment, ...) :
  <connection>:7: unexpected END_OF_INPUT '
'
```

The example for `GetPubMedByID()` returns an warning I don't understand:

```
In ProcessDate(bib[["year"]], bib[["month"]]) :
```

The example for `ReadCrossRef()` returns an error for this part of the example: `ReadCrossRef(filter = list(prefix = "10.5555"), limit = 5, min.relevance = 0)`,

```
server error for doi “http://dx.doi.org/10.5555/cs3112-testdoi-4”, you may want to try again.
```

and part: `ReadCrossRef(query = 'rj carroll measurement error', limit = 2, sort = "relevance", min.relevance = 80, use.old.api = TRUE)`

```
server error for doi “http://dx.doi.org/10.1002/0470011815.b2a03082”, you may want to try again.
```

I didn't see examples for `head()`, `PrintBibliography()`, `NoCite()`, `ReadPDFs()` or `tail()`.

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
Like Carl, I see references in the text but no reference list literally in the `.md` file. Not sure if this matters to the editors.

---

This is my first time doing a package review for rOpenSci, but I think this package looks great. RefManageR provides a variety of helpful functions for managing and citing bibliographic entries in RMarkdown. I was not previously aware of the package, but I will certainly be using it in my own work going forward. All the automated checks passed on my system, and the documentation is clear with thorough examples. As noted above, I ran in to a few warnings I didn't understand, but no errors or bugs.
