### 2017-07-18 #119 @mwmclean
Thank you for the helpful comments @AmeliaMN.  I think I've managed to address the issues you raised with some commits today up to [this one](https://github.com/mwmclean/RefManageR/commit/fa2aa45bdac9b8e9610c09b5ffc38e60e4b513ca).

- Regarding vignettes:
  - I removed the plot from the `"TestHtml"` vignette.
  - I feel the `"manual"` vignette already covers all three areas mentioned above in detail, with Section 3 of that vignette covering the external querying portion of the package.
- Regarding documenation:
  - I corrected the documentation "E-Utilities" typo mentioned
  - I added `httr::` whenever `http_error` is called in the documentation examples
  - The warning message in the `GetBibEntryWithDOI` example was due to not decoding the URLs in the returned BibTeX and has been corrected
  -  The `GetPubMedByID` example warning was fixed
  - I changed the two mentioned `ReadCrossRef` examples so they would not result in those messages about server errors and also changed the message to be more informative about what was happening
  - I added examples to the documentation for `head`, `PrintBibliography`, `NoCite`, `ReadPDFs`, and `tail`
