### 2017-07-23 #119 @noamross
Thanks for your reviews @cboettig and @AmeliaMN, and your follow-ups @mwmclean.  Let me just jump to ask @mwmclean to please let us know when you think your updates have addressed the all the reviewer comments thus far (including commenting here on big-picture stuff).  Then we can take a second look at all the changes in aggregate.

Also, your `paper.md` + `paper.bib` look fine to me.  JOSS will let us know if there's any issue in compiling when we transfer over.
