### 2017-07-25 #119 @mwmclean
In response to the `goodpractice::gp` checks, I have reduced the cyclomatic complexity of the flagged functions, except for the three I mentioned which I feel are false positives.  I have now also removed the check NOTE about foreign function call and also the import of `bibtex` thanks to that package's recent updates.  Thus, I feel I have finished responding to the initial checks.

Regarding the comments of @cboettig
* I have [now](https://github.com/mwmclean/RefManageR/commit/4dcbfd8bbad5d495f2f58301a358a280fe0142b5) removed all the `try` calls wrapping my API queries and instead use `httr::http_error` and `httr::stop_for_status`
* I think the idea of splitting the package into three separate packages for each of the mentioned "buckets" is interesting; I will open an issue for it, but I don't think I'll be able to get to this in the short term
* Some thoughts regarding `BibLaTeX` not being ideal:
  1) I don't believe the format returned by the various APIs considered in the package is standardized to the point where we can say there is one clear ideal choice, as it seems each has to be parsed on a case by case basis.  For example, the software case you mention for `CrossRef`/`DataCite` won't carry over to `Zotero` as far as I know
  2) There are several instances in the package where additional information returned by the API query is stored in a custom field (or e.g. the "note" field), which the user can easily access/manipulate (bucket 1) or format/print using a custom style they create (bucket 2)
  3) For storing bibliographic information for use with `LaTeX`, `BibLaTeX` seems to be the gold standard and because `LaTeX` is so widely used by statisticians and data scientists, I feel it is a good choice
  4) Since `BibTeX` is still more widely used by academic publishers, I provide functions to convert from `BibLaTeX` to `BibTeX` format and these are explained in the documentation and vignette

@noamross, let me know if I've missed anything, but I believe I've now addressed all the comments thus far.
