### 2017-07-25 mwmclean/RefManageR#34 @mwmclean
Split functionality/modules into separate packages
See [here](https://github.com/ropensci/onboarding/issues/119#issuecomment-311459131)
