### 2017-07-25 #119 @noamross
Thank you, @mwmclean. @cboettig and @AmeliaMN, please let us know if the updates have addressed all of your concerns for your reviews. If so, say so below and check off remaining boxes on your reviews. If not, please give details.
