### 2017-07-25 #119 @cboettig
Yup, 👍 , I think this is ready to go and I think i've checked all my boxes above.
