### 2017-07-26 #119 @AmeliaMN
Looking good! I'm still confused about the Entrez/E-Utilities thing, but I think that's just my lack of familiarity with NCBI.

Thanks for adding those additional examples. The only thing I can see now is the example for `ReadPDFs` isn't working because
```
Error in ReadPDFs(path) : poppler does not seem to be installed
```
