### 2017-07-26 #119 @noamross
I believe it's in the help file for `ReadPDFs`(), but may I suggest amending the error message so it reads
```
Poppler does not seem to be installed.
ReadPDFs requires that the `pdfinfo` utility from Poppler PDF be installed.
See http://poppler.freedesktop.org/
```
