### 2017-07-27 #119 @mwmclean
@noamross, thanks for the suggestion; this is [now](https://github.com/mwmclean/RefManageR/commit/d699094853f68abd13e1267ff5f9dc97add464eb) done.
