### 2017-07-27 #119 @noamross
Approved! Thanks @mwmclean for submitting and @cboettig and @AmeliaMN for your reviews!

To-dos:
```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
```
- [ ] Add the (brand new!) "Peer-reviewed" status badge to your README (in the future people will add this on submittal; it updates through the stages of review):
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)
- [ ] Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). Let us know if you are interested, and if so @stefaniebutland will be in touch about it when she's back from vacation late next week.
