### 2017-07-27 #119 @mwmclean
Thank you very much @noamross, @AmeliaMN, and @cboettig for helping improve the package!!  All the To-dos should be done now.
