### 2017-08-03 #119 @mwmclean
@noamross do you still need to turn on coverage services for the transferred repo?  I've tried setting up both Codecov and Coveralls with no luck.
