### 2017-08-03 #119 @noamross
There's an issue with permissions in the transfer to Codecov that we're figuring out that has held up the past couple of migrations.  Working on it!  Right now it looks like your .travis.yml sends to Codecov (which is not running), but your current badge is pointing to Coveralls.  I'll let you know when when have the codecov migration working.
