### 2017-08-03 #119 @mwmclean
I switched back to Coveralls. Not sure how long it takes to update, but coveralls.io is still saying "this repo has not received any jobs yet".
