### 2017-08-15 66a88ff
update for JOSS acceptance
