### 2017-08-17 40e5d5c
add image files for README since CMD check didnt like ROpenSci links
