+++ README.md
@@ -2,10 +2,10 @@ RefManageR
 ========
 [![](https://travis-ci.org/ropensci/RefManageR.svg?branch=master)](https://travis-ci.org/ropensci/RefManageR/)
 [![AppVeyor Build Status](http://ci.appveyor.com/api/projects/status/github/ropensci/RefManageR?branch=master&svg=true)](http://ci.appveyor.com/project/ropensci/RefManageR)
-[![Coverage Status](https://coveralls.io/repos/github/ropensci/RefManageR/badge.svg?branch=master)](https://coveralls.io/github/ropensci/RefManageR?branch=master)
+[![Coverage Status](https://s3.amazonaws.com/assets.coveralls.io/badges/coveralls_72.svg)](https://coveralls.io/github/mwmclean/RefManageR?branch=master)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/RefManageR)](https://cran.r-project.org/package=RefManageR)
 [![CRAN_Download_Badge](http://cranlogs.r-pkg.org/badges/RefManageR)](https://cran.r-project.org/package=RefManageR)
-[![](https://ropensci.org/badges/120_status.svg)](https://github.com/ropensci/onboarding/issues/120)
+[![](./inst/doc/120_status.svg)](https://github.com/ropensci/onboarding/issues/119)
 [![](http://joss.theoj.org/papers/10.21105/joss.00338/status.svg)](http://joss.theoj.org/papers/10.21105/joss.00338)
 
 `RefManageR` provides tools for importing and working with
@@ -31,4 +31,4 @@ To install the latest version from `GitHub`:
 install.packages("devtools")
 devtools::install_github("ropensci/RefManageR")
 ```
-[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
+[![ropensci_footer](./inst/doc/ropensci_footer.png)](https://ropensci.org)
