### 2017-08-17 dac1614
fix link to ROpenSci onboarding thread
