### 2017-08-23 ee05b04
cleanup temp file after GetCrossRefBibTeX
