### 2017-08-25 #119 @sckott
hi @mwmclean - coveralls is turned on now for this repo, can you try again?
