### 2017-08-25 #119 @mwmclean
@sckott I restarted a build on travis just after I saw your message.  Still saying no builds on [coveralls.io](https://coveralls.io/github/ropensci).
