### 2017-08-25 #119 @sckott
So i'm not that familiar with coveralls for R pkgs. It's turned on for your repo, but do you need a Coveralls key on Travis as an env var?  Used to need one for codecov, but now just connects automatically.
