### 2017-08-25 af1f1cd
try adding repo token for coveralls
