+++ .coveralls.yml
@@ -1,0 +1,1 @@
+repo_token: HstWDfCsRRi0oG0VVZFlEK4aXwWwh0t7R
