### 2017-08-28 #119 @mwmclean
Looks like something weird with my package or `covr`.  One of my tests is only failing during `covr::coveralls(type = "all")` even though the build is successful.
