### 2017-08-28 #119 @sckott
don't know @jimhester ? any thoughts?, see https://travis-ci.org/ropensci/RefManageR/builds/268404448#L2305
