### 2017-08-28 #119 @jimhester
To debug this you should run

```r
Rscript -e 'covr::package_coverage(type = "all", quiet = FALSE, clean = FALSE)'
```

Then you can inspect the package and libraries that are created locally.

When I do this I get the following output.

```
Error: identical(names(y), names) is not TRUE
Execution halted
```
