### 2017-08-30 #119 @stefaniebutland
Hello @mwmclean. Following up to ask if you are interested in writing a blog post about your package: either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/).
