### 2017-09-05 189bfd3
fix file removal when user supplies DOI, closes #35
