### 2017-10-02 #119 @stefaniebutland
ok with me to close re: blog post. @mwmclean please ping me if and when you want to do one
