### 2017-10-18 #119 @mwmclean
Sorry, did we need my input on the closing question?  I still need to track down my issue with `covr`, but I'm okay to close. I will be sure to message if I get the time to write a blog post.
